+++ services/inventory-service/src/main/java/com/example/inventory/controller/StoreController.java
@@ -1,6 +1,7 @@
 package com.example.inventory.controller;
 
-import com.example.inventory.dto.StoreDto;
+import com.example.inventory.dto.request.StoreRequest;
+import com.example.inventory.dto.response.StoreResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.ResponseEntity;
@@ -12,18 +13,18 @@ import java.util.UUID;
 public interface StoreController {
 
     @Operation(description = "가게 목록")
-    public ResponseEntity<List<StoreDto.Response>> getStores();
+    public ResponseEntity<List<StoreResponse>> getStores();
 
     @Operation(description = "가게 id 검색, 상세조회")
-    public ResponseEntity<StoreDto.Response> getStoreById(String id);
+    public ResponseEntity<StoreResponse> getStoreById(String id);
 
     @Operation(description = "가게 등록")
-    public ResponseEntity<UUID> addStore(StoreDto.Request request);
+    public ResponseEntity<UUID> addStore(StoreRequest request);
 
     @Operation(description = "가게 수정")
-    public ResponseEntity<StoreDto.Response> updateStore(String id, StoreDto.Request request);
+    public ResponseEntity<StoreResponse> updateStore(String id, StoreRequest request);
 
     @Operation(description = "가게 삭제")
-    public ResponseEntity<StoreDto.Response> deleteStore(String id);
+    public ResponseEntity<StoreResponse> deleteStore(String id);
 
 }
