+++ services/inventory-service/src/main/java/com/example/inventory/controller/StoreControllerImpl.java
@@ -1,6 +1,7 @@
 package com.example.inventory.controller;
 
-import com.example.inventory.dto.StoreDto;
+import com.example.inventory.dto.request.StoreRequest;
+import com.example.inventory.dto.response.StoreResponse;
 import com.example.inventory.service.store.StoreService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -20,33 +21,33 @@ public class StoreControllerImpl implements StoreController {
 
     @GetMapping
     @Override
-    public ResponseEntity<List<StoreDto.Response>> getStores() {
-        List<StoreDto.Response> response = storeService.findAll();
+    public ResponseEntity<List<StoreResponse>> getStores() {
+        List<StoreResponse> response = storeService.findAll();
         return ResponseEntity.ok(response);
     }
 
     @GetMapping("{id}")
     @Override
-    public ResponseEntity<StoreDto.Response> getStoreById(@PathVariable String id) {
+    public ResponseEntity<StoreResponse> getStoreById(@PathVariable String id) {
         return ResponseEntity.ok(storeService.findById(id));
     }
 
     @PostMapping
     @Override
-    public ResponseEntity<UUID> addStore(@RequestBody StoreDto.Request request) {
+    public ResponseEntity<UUID> addStore(@RequestBody StoreRequest request) {
         UUID uuid = storeService.create(request);
         return ResponseEntity.ok(uuid);
     }
 
     @PutMapping("{id}")
     @Override
-    public ResponseEntity<StoreDto.Response> updateStore(@PathVariable String id, @RequestBody StoreDto.Request request) {
+    public ResponseEntity<StoreResponse> updateStore(@PathVariable String id, @RequestBody StoreRequest request) {
         return ResponseEntity.ok(storeService.update(id, request));
     }
 
     @DeleteMapping("{id}")
     @Override
-    public ResponseEntity<StoreDto.Response> deleteStore(@PathVariable String id) {
+    public ResponseEntity<StoreResponse> deleteStore(@PathVariable String id) {
         storeService.delete(id);
         return ResponseEntity.noContent().build(); // 204
     }
